PrimaryButtonComponentを作成 #18
