+++ client/src/components/Dashboard/Top.js
@@ -1,5 +1,4 @@
 import React, { useState } from "react";
-import { makeStyles } from "@material-ui/core/styles";
 import Button from "@material-ui/core/Button";
 import PersonOutlineIcon from "@material-ui/icons/PersonOutline";
 import MailOutlineIcon from "@material-ui/icons/MailOutline";
@@ -7,27 +6,13 @@ import LockOutlinedIcon from "@material-ui/icons/LockOutlined";
 import { FontAwesomeIcon } from "@fortawesome/react-fontawesome";
 import TextInputWithIcon from "../Form/textInputWithIcon";
 import Navbar from "../Layout/Navbar";
-
-const useStyles = makeStyles(theme => ({
-  button: {
-    width: "120px",
-    height: "40px",
-    backgroundColor: "#172b4d",
-    color: "#FFFFFF",
-    boxShadow: "none",
-    "&:hover": {
-      backgroundColor: "#172b4d",
-      boxShadow: "none"
-    }
-  }
-}));
+import PrimaryButton from "../Button/PrimaryButton";
 
 const personIcon = <PersonOutlineIcon className="icon" />;
 const mailIcon = <MailOutlineIcon className="icon" />;
 const pwIcon = <LockOutlinedIcon className="icon" />;
 
 const Top = ({ headers, user, loggedIn, signup }) => {
-  const classes = useStyles();
   const [values, setState] = useState({
     name: "",
     email: "",
@@ -97,13 +82,7 @@ const Top = ({ headers, user, loggedIn, signup }) => {
                   handleChange={handleChange}
                 />
                 <div className="center">
-                  <Button
-                    type="submit"
-                    className={classes.button}
-                    variant="contained"
-                  >
-                    Sign up
-                  </Button>
+                  <PrimaryButton value="sign up" />
                 </div>
               </form>
             </div>
